fix(compat/flattenDepth): move flattening logic from flatten to flattenDepth
