+++ docs/ja/compat/reference/array/flatten.md
@@ -11,14 +11,14 @@
 配列を1段階平坦化します。
 
 ```typescript
-const result = flatten(array, depth);
+const result = flatten(array);
 ```
 
 ## 使用法
 
-### `flatten(value, depth)`
+### `flatten(array)`
 
-ネストされた配列を指定された深さだけ平坦化します。デフォルトでは1段階のみ平坦化し、ArgumentsオブジェクトやSymbol.isConcatSpreadableを持つオブジェクトもサポートします。
+ネストされた配列を1段階平坦化します。
 
 ```typescript
 import { flatten } from 'es-toolkit/compat';
@@ -27,19 +27,20 @@ import { flatten } from 'es-toolkit/compat';
 flatten([1, [2, [3, [4]], 5]]);
 // 結果: [1, 2, [3, [4]], 5]
 
-// 深さを指定
-flatten([1, [2, [3, [4]], 5]], 2);
-// 結果: [1, 2, 3, [4], 5]
-
 // Argumentsオブジェクトのサポート
 function example() {
   return flatten(arguments);
 }
 example(1, [2, 3], [[4]]);
 // 結果: [1, 2, 3, [4]]
+
+// Symbol.isConcatSpreadableを持つオブジェクトのサポート
+const spreadable = { 0: 'a', 1: 'b', length: 2, [Symbol.isConcatSpreadable]: true };
+flatten([1, spreadable, 3]);
+// 結果: [1, 'a', 'b', 3]
 ```
 
-空の配列やnull、undefinedは空の配列を返します。
+`null`や`undefined`は空の配列として処理されます。
 
 ```typescript
 import { flatten } from 'es-toolkit/compat';
@@ -49,20 +50,9 @@ flatten(undefined); // []
 flatten([]); // []
 ```
 
-Symbol.isConcatSpreadableを持つオブジェクトも配列のように平坦化されます。
-
-```typescript
-import { flatten } from 'es-toolkit/compat';
-
-const spreadable = { 0: 'a', 1: 'b', length: 2, [Symbol.isConcatSpreadable]: true };
-flatten([1, spreadable, 3]);
-// 結果: [1, 'a', 'b', 3]
-```
-
 #### パラメータ
 
-- `value` (`ArrayLike<T> | null | undefined`): 平坦化する配列です。
-- `depth` (`number`, オプション): 平坦化する最大深さです。デフォルトは`1`です。
+- `array` (`ArrayLike<T> | null | undefined`): 平坦化する配列です。
 
 #### 戻り値
